# wait_us optimization