--- a/hal/mbed_us_ticker_api.c
+++ b/hal/mbed_us_ticker_api.c
@@ -16,6 +16,7 @@
  */
 
 #include <stddef.h>
+#include "platform/mbed_atomic.h"
 #include "hal/us_ticker_api.h"
 
 #if DEVICE_USTICKER
@@ -24,15 +25,54 @@ static ticker_event_queue_t events = { 0 };
 
 static ticker_irq_handler_type irq_handler = ticker_irq_handler;
 
+#if MBED_CONF_TARGET_INIT_US_TICKER_AT_BOOT
+
+// If we are initializing at boot, we want the timer to be
+// always-on, so we block any attempt to free it. We do need
+// to pass through init(), as that needs to reset pending
+// interrupts.
+static void block_us_ticker_free()
+{
+}
+
+#else // MBED_CONF_TARGET_INIT_US_TICKER_AT_BOOT
+
+bool _us_ticker_initialized;
+
+// If we are not initializing at boot, we want to track
+// whether the timer has been initialized. This permits
+// a fast path for wait_us.
+static void note_us_ticker_init()
+{
+    us_ticker_init();
+    core_util_atomic_store_bool(&_us_ticker_initialized, true);
+}
+
+static void note_us_ticker_free()
+{
+    core_util_atomic_store_bool(&_us_ticker_initialized, false);
+    us_ticker_free();
+}
+
+#endif // MBED_CONF_TARGET_INIT_US_TICKER_AT_BOOT
+
 static const ticker_interface_t us_interface = {
+#if MBED_CONF_TARGET_INIT_US_TICKER_AT_BOOT
     .init = us_ticker_init,
+#else
+    .init = note_us_ticker_init,
+#endif
     .read = us_ticker_read,
     .disable_interrupt = us_ticker_disable_interrupt,
     .clear_interrupt = us_ticker_clear_interrupt,
     .set_interrupt = us_ticker_set_interrupt,
     .fire_interrupt = us_ticker_fire_interrupt,
     .get_info = us_ticker_get_info,
-    .free = us_ticker_free,
+#if MBED_CONF_TARGET_INIT_US_TICKER_AT_BOOT
+    .free = block_us_ticker_free,
+#else
+    .free = note_us_ticker_free,
+#endif
     .runs_in_deep_sleep = false,
 };