--- a/platform/mbed_toolchain.h
+++ b/platform/mbed_toolchain.h
@@ -277,11 +277,11 @@
  */
 #ifndef MBED_FORCEINLINE
 #if defined(__GNUC__) || defined(__clang__) || defined(__CC_ARM)
-#define MBED_FORCEINLINE static inline __attribute__((always_inline))
+#define MBED_FORCEINLINE inline __attribute__((always_inline))
 #elif defined(__ICCARM__)
-#define MBED_FORCEINLINE _Pragma("inline=forced") static
+#define MBED_FORCEINLINE _Pragma("inline=forced")
 #else
-#define MBED_FORCEINLINE static inline
+#define MBED_FORCEINLINE inline
 #endif
 #endif