--- a/targets/TARGET_Freescale/TARGET_MCUXpresso_MCUS/TARGET_KL27Z/us_ticker_defines.h
+++ b/targets/TARGET_Freescale/TARGET_MCUXpresso_MCUS/TARGET_KL27Z/us_ticker_defines.h
@@ -1,0 +1,30 @@
+/* mbed Microcontroller Library
+ * Copyright (c) 2006-2019 ARM Limited
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#ifndef _FSL_US_TICKER_DEFINES_H_
+#define _FSL_US_TICKER_DEFINES_H_
+
+#include "fsl_pit.h"
+
+#define US_TICKER_PERIOD_NUM    1
+#define US_TICKER_PERIOD_DEN    1
+
+#define US_TICKER_MASK          0xFFFFFFFF
+
+/* Macro-optimised form of us_ticker_read */
+#define us_ticker_read() (~(PIT_GetCurrentTimerCount(PIT, kPIT_Chnl_1)))
+
+#endif /* _FSL_US_TICKER_DEFINES_H_ */