--- a/targets/TARGET_Freescale/TARGET_MCUXpresso_MCUS/TARGET_MCU_K64F/us_ticker.c
+++ b/targets/TARGET_Freescale/TARGET_MCUXpresso_MCUS/TARGET_MCU_K64F/us_ticker.c
@@ -15,6 +15,7 @@
  */
 #include <stddef.h>
 #include "us_ticker_api.h"
+#include "us_ticker_defines.h"
 #include "PeripheralNames.h"
 #include "fsl_pit.h"
 #include "fsl_clock_config.h"
@@ -82,9 +83,9 @@ void us_ticker_init(void)
  *
  * @return The current timer's counter value in ticks
  */
-uint32_t us_ticker_read()
+uint32_t (us_ticker_read)()
 {
-    return ~(PIT_GetCurrentTimerCount(PIT, kPIT_Chnl_1));
+    return us_ticker_read();
 }
 
 /** Disable us ticker interrupt