--- a/targets/TARGET_Freescale/TARGET_MCUXpresso_MCUS/api/objects.h
+++ b/targets/TARGET_Freescale/TARGET_MCUXpresso_MCUS/api/objects.h
@@ -103,6 +103,8 @@ struct qspi_s {
 
 #include "gpio_object.h"
 
+#include "us_ticker_defines.h"
+
 #ifdef __cplusplus
 }
 #endif