… targets/TARGET_STM/us_ticker_defines.h
… targets/targets.json