### Description

As the timer code became more generic, coping with initialization on demand, and variable width and speed us_ticker_api implementations, `wait_us` has gradually gotten slower and slower.

Some platforms have reportedly seen overhead of `wait_us()` increase from 10µs to 30µs. These changes should fully reverse that drop, and even make it better than ever.

Add fast paths for platforms that provide compile-time information about `us_ticker`. Speed and code size is improved further if:

* Timer has >= 2^32 microsecond range, or better still is 32-bit 1MHz.
* Platform implements `us_ticker_read()` as a macro
* Timer is initialised at boot, rather than first use

The latter initialisation option is the default for STM, as this has always been the case.

PR adds support for the optimization for Freescale MCUXpresso family and all STM devices.

Testing on K64F, a tight `{ led = !led; wait_us(10); }` loop does achieve 10µs transitions with this change. Without, it was 16µs.

Information provided by targets is only currently used to optimize `wait_us`, not any other use of us_ticker via `Ticker` et al, but the information could be used to optimize those in future.

### Pull request type

    [ ] Fix
    [X] Refactor
    [ ] Target update
    [ ] Functionality change
    [ ] Docs update
    [ ] Test update
    [ ] Breaking change

### Reviewers

@bulislaw 

### Release Notes

* `wait_us` has been optimized for certain platforms - see `us_ticker_api.h` for the details of enabling optimizations on other targets. It can be optimized further if the option `target.init-us-ticker-at-boot` is enabled.