@ciarmcom 2019-05-17
@kjbracey-arm, thank you for your changes.
@maclobdell @bulislaw @MarceloSalazar @ARMmbed/mbed-os-core @ARMmbed/mbed-os-test @ARMmbed/mbed-os-maintainers @ARMmbed/mbed-os-hal please review.