@jeromecoutant 2019-05-20
Hi
After some tests, it seems that "Microsecond ticker init is safe to call repeatedly" test from tests-mbed_hal-common_tickers becomes FAIL for all STM32 families