@kjbracey-arm 2019-05-20
@jeromecoutant, this should have fixed the `init()` test issue.

Our tests have shown that this does greatly improve performance on STM32L151, which we're particularly targeting, but apparently performance is still below Mbed OS 5.6.1 with this change - ~12µs overhead, when 5.6.1 was ~9µs. The generated code for `wait_us` is optimal, so any delay is now down to the timer itself, or general system load.

Is there any reason in the STM target code that the timer would be slower now than in 5.6.1 on that device?  (We're using 5.6.1 as a comparator, as that's the last time the `wait_us` code was small and tight with just a straight call to the target's `us_ticker_read`).

Would running the timer faster than 1MHz improve read time?