@kjbracey-arm 2019-05-20
> Looks good, is there a point of doing something similar for LP ticker?

Potentially, yes. Less necessary for speed of lpticker itself, but would open the door to replacing the whole ticker_data/api core with something templatey - get as much stuff figured out at compile time as possible.

Total code size of  two template specialisations for lpticker and usticker might come out not too dissimilar to the generic code.

(The choice to use numerator/denominator ratios here was influenced by C++11 `std::ratio` and `std::duration` - might want to use that as part of the framework).