@kjbracey-arm 2019-05-20
> The change looks good however it is confusing to have macros and functions with the same name.

Somewhat agree, but I feel our hands are partially tied by binary compatibility, in particular for `wait_us`. I bet there are wifi drivers containing us timing calls :/. I don't think we can afford for `wait_us` or `us_ticker_read` to not have a real definition, nor can we change the prototype, I think. 

The scheme here is "conventional" in that it's what the standard C library officially endorses - any standard library function can be implemented as a macro, as long as there is a real function too that you can take the address of.

Now, I've only just realised that there is precedent in the HAL for inlining - `gpio_read` and `gpio_write` for STM. That has `void gpio_write()` in `gpio_api.h`, but STM's `gpio_objects.h` defines `static inline void gpio_write()`.  (And `gpio_api.h` always includes `device.h`, so the static definition is always visible, and there is no external definition).

Personally I find that confusing as well - it's jarring to me that `void gpio_write()` is a valid declaration of a static function....

> the default implementation of us_ticker_read as weak and override it as a static inline function in header files of platform supporting this optimisation

That doesn't quite add up - a static function wouldn't override a weak external function, it would just hide it.

There is no "default" implementation of `us_ticker_read` - it's the core HAL function.  I guess defining the real function as "invoke the HAL macro" could be provided by the core, when there is a macro? But again, I feel this is safest for binary compatibility - keeping the HAL providing an external `us_ticker_read` always. 

Alternatively, if we did act like `gpio_write` and have a static inline, we would find it hard to provide a real external `us_ticker_read` for backwards compatibility - we'd have to suppress the static definition in whichever file provided the external one. That's harder than suppressing the macro with the extra `()` around the function identifier.




Which means for STM there actually is no external `gpio_read` definition, only static inline definitions.

But I'm wary of revoking