@pan- 2019-05-20
Thank you for that insightful answer @kjbracey-arm; I it wasn't really clever to suggest an override of a global symbol with a static inline one 😬 . I wonder if there's some compiler magic we can use to have a global function defined inline as we want the definition to be available for all translation unit.