@kjbracey-arm 2019-05-21
> I wonder if there's some compiler magic we can use to have a global function defined inline as we want the definition to be available for all translation unit.

C99 non-static inline kind of works like that, and would preserve binary compatibility. You can have an inline definition in your header file, and there is exactly one real external definition in one translation unit - all other translation units either use the inline definition or call/address that external definition.

There is a minor potential source compatibility glitch that was bothering me, if people were declaring the function themselves, which is why I avoided it, but I've realised that the macro form would also fail in that same situation.

I'll change to non-static inline, as it does correspond to `gpio_write`, and is a bit clearer, if you agree.