@pan- 2019-05-21
What was the source compatibility glitch that bothered you ? Other than that I'm delighted if we use the inline function solution.