@kjbracey-arm 2019-05-21
> What was the source compatibility glitch that bothered you ?

If there's an inline definition in the header, then this C file would produce an extra duplicate external definition, causing a link error:

    #include "mbed.h"
    extern uint32_t us_ticker_read();

(Not sure if the `extern` is needed to provoke that, or just lack of `inline`; standard isn't clear)

But if it was a macro, then the above would also fail to compile as it would substitute the macro - that's why `us_ticker_api.h` had its declaration adjusted to `uint32_t (us_ticker_read)()`.

So extra user declarations break either way.