@kjbracey-arm 2019-05-21
Ah, another stumbling block. K64F's `us_ticker_read` would want to call its `PIT_GetCurrentTimerCount`, which is `static inline`.

Non-static inline functions can't call static functions.

We have quite a lot of `static inline` in various places, substituting for macros, but it's quite hard to mix the two types of `inline`. :(