@mmahadevan108 2019-05-28
I am seeing the below error when running the usticker test on K66F

Scan: us_ticker
Compile [100.0%]: main.cpp
[Error] main.cpp@39,89: '_ticker_info' was not declared in this scope
[Error] main.cpp@40,86: 'TEST_ASSERT_UINT32_EQUAL' was not declared in this scope
Memory map breakdown for built projects (values in Bytes):
| name | target | toolchain | static_ram | total_flash |
|------|--------|-----------|------------|-------------|


Build successes:
  * K66F::GCC_ARM::MBED-BUILD


Build failures:
  * K66F::GCC_ARM::MBED-OS-TESTS-MBED_HAL-US_TICKER
        Building project us_ticker (K66F, GCC_ARM)
        Scan: GCC_ARM
        Scan: us_ticker
        Compile [100.0%]: main.cpp
        [Error] main.cpp@39,89: '_ticker_info' was not declared in this scope
        [Error] main.cpp@40,86: 'TEST_ASSERT_UINT32_EQUAL' was not declared in this scope

[mbed] ERROR: "c:\python27\python.exe" returned error.
       Code: 1
       Path: "C:\mymbed"
       Command: "c:\python27\python.exe -u C:\mymbed\mbed-os\tools\test.py -D MBED_TEST_MODE -t GCC_ARM -m K66F -c --sou
rce . --build .\BUILD\tests\K66F\GCC_ARM --test-spec .\BUILD\tests\K66F\GCC_ARM\test_spec.json --build-data .\BUILD\test
s\K66F\GCC_ARM\build_data.json -n mbed-os-tests-mbed_hal-us* --greentea"
       Tip: You could retry the last command with "-v" flag for verbose output
---