@adbridge 2019-06-18
@pan- @bulislaw  could we have a re-review please ?